requirements.txt
bcrypt==1.1.1
cffi==0.9.2
Django==1.7.7
MySQL-python==1.2.5
Pillow==2.8.1
pip==6.0.8
pycparser==2.10
PyJWT==1.0.1
South==1.0.2
